devsocial_api/requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.3
python-dotenv
django-storages
boto3
bcrypt